% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{get_nrel_api_email}
\alias{get_nrel_api_email}
\title{Retrieve saved API email address}
\usage{
get_nrel_api_email(invisible = TRUE)
}
\arguments{
\item{invisible}{}
}
\description{
Retrieve saved API email address
}
