% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{get_nrel_api_key}
\alias{get_nrel_api_key}
\title{Retrieve saved API key}
\usage{
get_nrel_api_key(invisible = TRUE)
}
\arguments{
\item{invisible}{Should the key be invisible (TRUE) or printed in the console (FALSE)?}
}
\value{
The key
}
\description{
Retrieve saved API key
}
\examples{
  get_nrel_api_key()
  get_nrel_api_key(F)
}
