% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{get_nrel_url}
\alias{get_nrel_url}
\title{Get NREL's API addresses by the database name}
\usage{
get_nrel_url(x)
}
\arguments{
\item{x}{a short name of the database}
}
\value{
character string with the URL address
}
\description{
Get NREL's API addresses by the database name
}
