% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{guess_nrel_grid}
\alias{guess_nrel_grid}
\title{Guess grid with locations of data-points in NREL's collection}
\usage{
guess_nrel_grid(
  gis_sf,
  dx = 0.01,
  dy = 0.01,
  q_atributes = "windspeed_100m",
  collection = "wtk",
  q_interval = 60,
  q_names = "2014",
  steps_max = 10,
  verbose = T
)
}
\arguments{
\item{gis_sf}{sf object (map) of the region to estimate a number of data-points in NREL's collection, and their coordinates (grid).}

\item{dx}{starting step along longitude}

\item{dy}{starting step along latitude}

\item{q_atributes}{attributes of the query (see \code{fetch_nrel_data})}

\item{collection}{name of NREL's data-collection (see \code{get_nrel_url})}

\item{q_interval}{interval parameter in the query}

\item{q_names}{names parameter in the query}

\item{steps_max}{maximum steps in evaluation of the grid along each axis, 10 by default}

\item{verbose}{logical, should the process be reported}
}
\value{
list with the grid in sf points format, grid parameters, fetched initial data-points used for the estimation of the grid.
}
\description{
Guess grid with locations of data-points in NREL's collection
}
