% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{nrel_fetch_coord}
\alias{nrel_fetch_coord}
\title{Fetch (request and download) data from NREL}
\usage{
nrel_fetch_coord(lon, lat, api_url = NULL, as = "raw", ...)
}
\arguments{
\item{lon}{vector with longitude coordinate(s)}

\item{lat}{vector with latitude coordinate(s)}

\item{api_url}{character string with the url-address to a particular dataset}

\item{as}{httr parameter}

\item{...}{attributes for the API query}
}
\description{
Fetch (request and download) data from NREL
}
