% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{nrel_get_url}
\alias{nrel_get_url}
\title{Get NREL's API addresses by the database name}
\usage{
nrel_get_url(x)
}
\arguments{
\item{x}{a short name of the database}
}
\value{
character string with the URL address
}
\description{
Get NREL's API addresses by the database name
}
