% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{nrel_read_meta}
\alias{nrel_read_meta}
\title{Extract meta-data from NREL's API response}
\usage{
nrel_read_meta(x)
}
\arguments{
\item{x}{raw response object returned by \code{httr::GET} function used in}
}
\description{
Extract meta-data from NREL's API response
}
