% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nrel_read_meta_files}
\alias{nrel_read_meta_files}
\title{Read meta-data from saved files with raw data}
\usage{
nrel_read_meta_files(
  ff,
  read_dir = NULL,
  workers = future::availableCores()\%/\%2
)
}
\arguments{
\item{ff}{character vector with file names (fill or base).}

\item{read_dir}{directory with files to read (if \code{ff} contains base names)}

\item{workers}{integer, number of workers (files to process in parallel)}
}
\description{
Read meta-data from saved files with raw data
}
