% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{nrel_read_responce}
\alias{nrel_read_responce}
\title{Extract content from NREL's API response}
\usage{
nrel_read_responce(x)
}
\arguments{
\item{x}{}
}
\description{
Extract content from NREL's API response
}
