% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nrel_read_save_data}
\alias{nrel_read_save_data}
\title{Read saved raw data from files and safe results in files by locations}
\usage{
nrel_read_save_data(
  meta,
  read_dir,
  save_dir,
  columns = NULL,
  col_mask = NULL,
  scale_round = 2,
  compress = 100,
  overwrite = TRUE,
  merge_files = TRUE,
  workers = max(future::availableCores()\%/\%2, 5)
)
}
\arguments{
\item{meta}{data.frame with IDs and files to process (output of \code{nrel_read_meta_files}, filtered for locations of interest).}

\item{read_dir}{path to the directory to files with raw data (output of \code{nrel_fetch_points} function).}

\item{save_dir}{path to the directory to save data.}

\item{columns}{columns of data to read from the data.}

\item{compress}{compression level, value in the range 0 to 100, an argument to \code{fst::write_fst} function.}

\item{overwrite}{logical, should files be overwritten if exist. If FALSE and the file already exists, a number will be added to the file name, indicating the version.}

\item{merge_files}{unused}

\item{workers}{integer, number of workers - files to process in parallel (affects memory use).}
}
\value{
saved files in \code{save_dir} with derived data from "raw" files.
}
\description{
Read saved raw data from files and safe results in files by locations
}
