% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrel_sitecounts_sf}
\alias{nrel_sitecounts_sf}
\title{Site counts in NREL collections.}
\format{
A simple features (\code{sf}) object with 22326 rows (1x1 cells) and 37 variables - collections with number of time series in each grid cell:
\describe{
\item{himawari7}{...}
\item{himawari-tmy}{...}
...
}
}
\usage{
nrel_sitecounts_sf
}
\description{
A dataset containing site-counts on 1x1 degree greed.
}
\keyword{datasets}
