% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_nrel_api_key}
\alias{set_nrel_api_key}
\title{Set API key and save it in a hidden file. The key will be retrieved with loading the package.}
\usage{
set_nrel_api_key(nrel.api.email, nrel.api.key)
}
\arguments{
\item{nrel.api.email}{API-registered email address}

\item{nrel.api.key}{the key received from NREL developer network.}
}
\value{
\verb{~/.nrel.dev.api} file with saved key and added to the environment options.
}
\description{
Set API key and save it in a hidden file. The key will be retrieved with loading the package.
}
\examples{
DO NOT RUN
set_nrel_api_key("YOUR_API_KEY")
}
