% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_data_from_grid}
\alias{fetch_data_from_grid}
\title{Download data from NREL using grid of points.}
\usage{
fetch_data_from_grid(
  grid_sf,
  cells = NULL,
  nmax = 100,
  plot_process = T,
  file_name = NULL,
  sleep_after_finish = FALSE
)
}
\arguments{
\item{grid_sf}{sf object with coordinates of points to download}

\item{cells}{integer sequence of row-numbers in \code{grid_sf} object}

\item{nmax}{integer maximum number of data-points to download}

\item{plot_process}{logical, should the process be visualized}

\item{file_name}{file name with path, if provided, the results will be saved in the file. Otherwise, results will be returned to global environment.}

\item{sleep_when_done}{logical, if TRUE, puts the computer to sleep when done}
}
\description{
Download data from NREL using grid of points.
}
