% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_nrel_data}
\alias{fetch_nrel_data}
\title{Title}
\usage{
fetch_nrel_data(lon, lat, api_url = NULL, as = "raw", ...)
}
\arguments{
\item{lon}{vector with longitude coordinate(s)}

\item{lat}{vector with latitude coordinate(s)}

\item{api_url}{character string with the API url}

\item{as}{httpr parameter}

\item{...}{other attributes for the API query}
}
\description{
Title
}
